--- model-1.1/调用示例.py
@@ -1,2 +1,0 @@
-from model import change_detection
-output_png_path = change_detection('ottawa_1.bmp', 'ottawa_2.bmp')
